pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research-domain"
version = "0.1.0"
description = "Library for managing research data: groups, campus, university, researchers, and papers."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Paulo Sérgio dos Santos Júnior", email = "paulossjunior@gmail.com" },
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "libbase @ git+https://github.com/The-Band-Solution/libbase.git@v0.1.0",
    "eo_lib @ git+https://github.com/The-Band-Solution/eo_lib.git@v0.3.2",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "black", "mypy"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/research_domain"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true
implicit_optional = true
disable_error_code = ["assignment"]

[[tool.mypy.overrides]]
module = "libbase.*"
ignore_missing_imports = true
